requirements.txt
fastapi==0.65.2
python-multipart==0.0.5
python-dotenv==0.17.1
sqlalchemy==1.4.22
alembic==1.6.5
psycopg2-binary==2.9.1
requests==2.26.0
beautifulsoup4==4.9.3
nltk==3.6.2
numpy==1.21.2
pandas==1.3.2
scikit-learn==0.24.2
joblib==1.0.1
pdfminer.six==20200726
python-docx==0.8.11
python-pptx==0.6.19
gunicorn==20.1.0
uvicorn==0.14.0
jinja2==3.0.1
aiofiles==0.7.0
pytest==6.2.4
pytest-asyncio==0.15.1
pytest-cov==2.12.1
coverage==5.5
mock==4.0.3
